requirements.txt
python-telegram-bot==20.8
psycopg2-binary>=2.9.9
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
PyJWT==2.10.1
python-multipart==0.0.20
